Downstream Application/testapp/backend/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
xgboost>=1.5.0
requests>=2.25.0
python-dotenv>=0.19.0
db-dtypes>=1.0.0
